+++ app/serializers/cosmere/character_serializer.rb
@@ -8,7 +8,7 @@ module Cosmere
                :movement, :recovery_die, :senses_range, :level, :abilities, :guide_step, :health, :focus, :investiture,
                :attribute_points, :skill_points, :deflect, :additional_skills, :tier, :ancestry, :cultures, :attacks, :money,
                :talent_points, :updated_at, :expertises, :custom_expertises, :features, :modified_abilities, :purpose, :obstacle,
-               :goals, :connections, :singer_forms, :singer_form, :names, :setting, :max_abilities
+               :goals, :connections, :singer_forms, :singer_form, :names, :setting, :max_abilities, :conditions_v2
 
     delegate :skills, :defense, :focus_max, :investiture_max, :load, :movement, :recovery_die, :senses_range, :max_abilities,
              :deflect, :tier, :attacks, :talent_points, :features, :health_max, :modified_abilities, :singer_forms,
@@ -16,7 +16,7 @@ module Cosmere
     delegate :data, to: :object
     delegate :attribute_points, :skill_points, :health, :focus, :investiture, :level, :guide_step, :additional_skills, :ancestry,
              :cultures, :expertises, :custom_expertises, :abilities, :purpose, :obstacle, :goals, :connections, :singer_form,
-             :setting, :money, to: :data
+             :setting, :money, :conditions_v2, to: :data
 
     def provider
       'cosmere'