+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - unified custom features
 - money management for cosmere/nimble characters
+- conditions management for cosmere characters
 
 ### Modified
 - leveling up safe mode